requirements.txt
albumentations==1.3.0
joblib==1.2.0
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
scikit_learn==1.2.2
streamlit==1.20.0
torch==2.0.0
torchvision==0.15.1
